+++ inst/plotlyjs.R
@@ -1,15 +1,17 @@
 library(httr)
 x <- GET('https://api.github.com/repos/plotly/plotly.js/releases/latest')
 zip <- content(x)$zipball_url
+# for a particular version:
+# zip <- "https://github.com/plotly/plotly.js/archive/v1.16.3.zip"
 tmp <- tempfile(fileext = ".zip")
 download.file(zip, tmp)
 unzip(tmp)
 # update the plotly.js bundle
-p <- Sys.glob("plotly-plotly.js*/dist/plotly.min.js")
+p <- Sys.glob("*plotly.js*/dist/plotly.min.js")
 file.copy(p, "inst/htmlwidgets/lib/plotlyjs/plotly-latest.min.js", overwrite = T)
-l <- Sys.glob("plotly-plotly.js*/LICENSE")
+l <- Sys.glob("*plotly.js*/LICENSE")
 file.copy(l, "inst/htmlwidgets/lib/plotlyjs/LICENSE", overwrite = T)
-unlink("plotly-plotly.js*", recursive = T)
+unlink("*plotly.js*", recursive = T)
 message("Manually update plotly.yaml with this version")
 basename(zip)